requirements.txt
google-genai>=0.5.0
python-dotenv>=1.0.1
pytest>=8.0.0
playwright>=1.40.0
pytest-playwright>=0.4.3
